Move cal_gedm to deepks_basic.cpp
