+++ source/Makefile.Objects
@@ -205,7 +205,6 @@ OBJS_DEEPKS=LCAO_deepks.o\
         LCAO_deepks_pdm.o\
         LCAO_deepks_phialpha.o\
         LCAO_deepks_interface.o\
-        cal_gedm.o\
         
 
 OBJS_ELECSTAT=elecstate.o\
